Move sourceCommit file and minor renaming
