+++ src/services/git.ts
@@ -3,13 +3,13 @@ import del from 'del';
 import { uniq, isEmpty } from 'lodash';
 import ora from 'ora';
 import { ValidConfigOptions } from '../options/options';
-import { Commit } from '../types/commitWithAssociatedPullRequests';
 import { HandledError } from './HandledError';
 import { execAsCallback, exec } from './child-process-promisified';
 import { getRepoOwnerPath, getRepoPath } from './env';
 import { stat } from './fs-promisified';
 import { getShortSha } from './github/commitFormatters';
 import { logger } from './logger';
+import { Commit } from './sourceCommit';
 
 async function folderExists(path: string): Promise<boolean> {
   try {
