+++ src/ui/getTargetBranches.ts
@@ -2,7 +2,7 @@ import { isEmpty, intersection } from 'lodash';
 import { ValidConfigOptions } from '../options/options';
 import { HandledError } from '../services/HandledError';
 import { promptForTargetBranches } from '../services/prompts';
-import { Commit } from '../types/commitWithAssociatedPullRequests';
+import { Commit } from '../services/sourceCommit';
 import { filterNil } from '../utils/filterEmpty';
 
 export function getTargetBranches(
